Capado permisos para admin

+++ app/Http/Controllers/AdminController.php
@@ -12,36 +12,62 @@ use Illuminate\Support\Facades\Mail;
 class AdminController extends Controller
 {
     public function users(){
-    $users = User::paginate();
-    return view('users',compact('users'));
+    if(session('tipo')=='admin'){
+        $users = User::paginate();
+        return view('users',compact('users'));
+    }
+    else{
+        return redirect('/');
+    }
+    
 
   }
 
     public function modificar($id){
-    $user = User::find($id);   
-    return view('modificar',compact('user'));
+    if(session('tipo')=='admin'){
+        $user = User::find($id);   
+        return view('modificar',compact('user'));
+    }
+    else{
+        return redirect('/');
+    }
 
   }
     
     public function update(Request $req){
-        $data=User::find($req->id);
-        $data->name=$req->name;
-        $data->email=$req->email;
-        $data->surname=$req->surname;
-        $data->save();
-        return redirect('users');
+        if(session('tipo')=='admin'){
+            $data=User::find($req->id);
+            $data->name=$req->name;
+            $data->email=$req->email;
+            $data->surname=$req->surname;
+            $data->save();
+            return redirect('users');
+        }
+        else{
+            return redirect('/');
+        }
     }
     
      public function delete($id){
+     if(session('tipo')=='admin'){
         $data = User::find($id);   
         $data->delete();
         return redirect('users');
+     }
+        else{
+            return redirect('/');
+        }
 
   }
     public function send_signup_mail($id){
-        $user=User::find($id);
-        $correo= new gestionSociosMailable($user->confirmation_code);
-        Mail::to($user->email)->send($correo);
-        return view('confirmacion_enviada');
+        if(session('tipo')=='admin'){
+            $user=User::find($id);
+            $correo= new gestionSociosMailable($user->confirmation_code);
+            Mail::to($user->email)->send($correo);
+            return view('confirmacion_enviada');
+        }
+        else{
+            return redirect('/');
+        }
     }
 }
